Rename singleton.py to state.py
The file is not about the "singleton" any more, it's about everything related to
the state.
